+++ src/appLoader.js
@@ -22,7 +22,8 @@ const path = require('path');
 // const app = require('connect')();
 const oas3Tools = require('oas3-tools');
 const logger = require('./log/logger');
-const OAuthHelper = require('./oauth/OAuthHelper');
+const AuthMiddleware = require('./middleware/AuthMiddleware');
+const SessionConfig = require('./oauth/SessionConfig');
 const HubCAService = require('./service/HubCAService');
 
 const db = require('./db/database');
@@ -50,7 +51,7 @@ exports.connect = async () => {
     openApiValidator: {
       validateSecurity: {
         handlers: {
-          OAuth2: Constants.OAUTH.AUTH_ENABLED ? OAuthHelper.createOAuth2Handler() : () => true,
+          OAuth2: Constants.OPENID.ENABLED ? AuthMiddleware.createOAuth2Handler() : () => true,
         }
       }
     }
@@ -97,6 +98,27 @@ exports.connect = async () => {
     cors(corsUtils.getCorsOptions),
     logger.createWinstonLogger()
   ];
+
+  // Add authentication middleware if enabled
+  if (Constants.OPENID.ENABLED) {
+    // Add session middleware for browser clients
+    middlewares.push(SessionConfig.createSessionMiddleware());
+
+    // Add authentication middleware for all clients
+    middlewares.push(AuthMiddleware.createAuthMiddleware());
+  } else {
+    middlewares.push((req, res, next) => {
+      const roles = req.headers['x-roles'];
+      if (roles) {
+        try {
+          req.user = { roles: JSON.parse(roles) };
+        } catch (e) {
+          console.log("Error getting roles from request header: ", e.message);
+        }
+      }
+      next();
+    });
+  }
 
   app.use(...middlewares);
   const stack = app._router.stack;